main_app_backend/user_server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
blinker==1.4
click==8.1.7
cryptography==3.4.8
distro==1.7.0
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
hiredis==2.2.3
httplib2==0.20.2
httptools==0.6.0
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pydantic_core==2.6.3
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==1.0.0
sqlmodel==0.0.8
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
wadllib==1.3.6
